% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{ag_expdetails}
\alias{ag_expdetails}
\title{Experimental details client wrapper}
\usage{
ag_expdetails(
  studyDbId = NULL,
  format = c("json", "list", "data.frame"),
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/002/r",
  endPoint = "/experiment/get.php",
  ...
)
}
\arguments{
\item{studyDbId}{character agronomic study identifier from AGROFIMS database}

\item{...}{inherents arguments from AgAPIClient class and Ag}
}
\description{
wrapper function for ragapi.ag_experiment. objects
}
