% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_wrappers.r
\name{ag_get_cropsite_expsiteId}
\alias{ag_get_cropsite_expsiteId}
\title{Wrapper for crop-site information GET method}
\usage{
ag_get_cropsite_expsiteId(
  expsiteDbId = NULL,
  format = c("json", "list", "data.frame"),
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0233/r",
  ...
)
}
\arguments{
\item{expsiteDbId}{character experiment site ID from AgroFIMS database}

\item{format}{data format: json, list and data.frame}

\item{serverURL}{URL of the agrofims server}

\item{version}{version of the call. By default version \code{0212}.}

\item{...}{additional parameters}
}
\description{
wrapper function for \code{objetc$ag_get_cropsite_expsiteId} method from \code{AgProjEntity} objects.
}
\examples{
\dontrun{
#Get crop-site description information by expsiteDbId (experiment site ID)
out <- ag_get_cropsite_expsiteId(expsiteDbId = 9, 
                                 format = "data.frame",
                                 serverURL = "https://research.cip.cgiar.org/agrofims/api/dev", 
                                 version ="/0233/r")
}
}
\author{
Omar Benites
}
