% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{ag_get_expdetails_studyId}
\alias{ag_get_expdetails_studyId}
\title{Wrapper for experimental details get method}
\usage{
ag_get_expdetails_studyId(
  studyDbId = NULL,
  format = c("json", "list", "data.frame"),
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0212/r",
  ...
)
}
\arguments{
\item{studyDbId}{character agronomic study identifier from AGROFIMS database}

\item{format}{data format: json, list and data.frame}

\item{serverURL}{URL of the agrofims server}

\item{version}{version of the call. By default version \code{0212}.}

\item{...}{additional parameters}
}
\description{
wrapper function for \code{objetc$ag_get_expdetails_studyId} from \code{AgExpDetails} objects.
}
\examples{
\dontrun{
#Get experimental details by studyId (agronomic study ID)
out <- ag_get_expdetails_studyId(studyDbId = 21, 
                                 format = "data.frame",
                                 serverURL = "https://research.cip.cgiar.org/agrofims/api/dev", 
                                 version ="/0212/r")
}
}
\author{
Omar Benites
}
