% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_wrappers.r
\name{ag_get_fertmea_expsiteId}
\alias{ag_get_fertmea_expsiteId}
\title{Wrapper for ag_get_fertmea_siteId  method}
\usage{
ag_get_fertmea_expsiteId(
  expsiteDbId = NULL,
  format = c("json", "list", "data.frame"),
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0291/r",
  ...
)
}
\arguments{
\item{expsiteDbId}{character experiment site ID from AgroFIMS database}

\item{format}{data format: json, list and data.frame}

\item{serverURL}{URL of the agrofims server}

\item{version}{version of the call. By default version \code{0253}.}

\item{...}{additional parameters}
}
\description{
wrapper function for \code{objetc$ag_get_fertmea_expsiteId} method from \code{AgFertilizerMea} objects.
}
\examples{
\dontrun{
Get fertilizer measurements, nutrients and products,  by expsiteDbId (experiment site ID)
out <- ag_get_fertmea_expsiteId(expsiteDbId = 7,
                                 format = "data.frame",
                                 serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                 version ="/0291/r")
}
}
\author{
Omar Benites
}
