% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_wrappers.r
\name{ag_get_fundagency_studyId}
\alias{ag_get_fundagency_studyId}
\title{Wrapper for funding agency get method}
\usage{
ag_get_fundagency_studyId(
  studyDbId = NULL,
  format = c("json", "list", "data.frame"),
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0253/r",
  ...
)
}
\arguments{
\item{studyDbId}{character agronomic study identifier from AGROFIMS database}

\item{format}{data format: json, list and data.frame}

\item{serverURL}{URL of the agrofims server}

\item{version}{version of the call. By default version \code{0212}.}

\item{...}{additional parameters}
}
\description{
wrapper function for \code{objetc$ag_get_fundagency_studyId} method from \code{AgFundAgency} objects.
}
\examples{
\dontrun{
#Get funding agencies information by studyId (agronomic study ID)
out <- ag_get_fundagency_studyId(studyDbId = 11,
                                 format = "data.frame",
                                 serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                 version ="/0253/r")
}
}
\author{
Omar Benites
}
