% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_wrappers.r
\name{ag_get_phenomea_expsiteId}
\alias{ag_get_phenomea_expsiteId}
\title{Wrapper for ag_get_phenomea_expsiteId method from AgPhenoMea class}
\usage{
ag_get_phenomea_expsiteId(
  expsiteDbId = NULL,
  format = c("json", "list", "data.frame"),
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0248/r",
  ...
)
}
\arguments{
\item{expsiteDbId}{character experiment site ID from AgroFIMS database}

\item{format}{data format: json, list and data.frame}

\item{serverURL}{URL of the agrofims server}

\item{version}{version of the call. By default version \code{0212}.}

\item{...}{additional parameters}
}
\description{
wrapper function for \code{objetc$ag_get_phenomea_expsiteId } method from \code{AgPhenoMea} objects.
}
\examples{
\dontrun{
Get crop measurements variables by expsiteDbId (experiment site ID)
library(ragapi)
out <- ag_get_phenomea_expsiteId(expsiteId = 6,
                                 format = "data.frame",
                                 serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                 version ="/0248/r")
}
}
\author{
Omar Benites
}
