% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgAPIClient.R
\name{AgAPIClient}
\alias{AgAPIClient}
\title{ApiClient class}
\description{
the AgAPI client provides a general class for all the AgAPI calls. It includes

Generic class for API client calls, all nbaR client classes
derive from this class

Specifies the public interface which any data accessing classes must implement.
}
\details{
AgAPI Client Class
}
\references{
Api url structure https://swagger.io/docs/specification/api-host-and-base-path/
}
\author{
Omar Benites
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{authentication}}{specifies the password, defaults to NULL}

\item{\code{token}}{specifies the password, defaults to NULL}

\item{\code{user_agent}}{specifies the password, defaults to NULL}

\item{\code{serverURL}}{specifies the password, defaults to 
http://research.cip.cgiar.org/agrofims/api/dev/}

\item{\code{version}}{specifies the version, defaults to 0051}

\item{\code{endPoint}}{specifies the end point value,   tableName/method.php}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{endPoint}}{specifies the end point value,   tableName/method.php}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AgAPIClient$new()}}
\item \href{#method-print}{\code{AgAPIClient$print()}}
\item \href{#method-call_api}{\code{AgAPIClient$call_api()}}
\item \href{#method-clone}{\code{AgAPIClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgAPIClient$new(
  user = NULL,
  password = NULL,
  authentication = NULL,
  token = NULL,
  user_agent = NULL,
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0135/r"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgAPIClient$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call_api"></a>}}
\if{latex}{\out{\hypertarget{method-call_api}{}}}
\subsection{Method \code{call_api()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgAPIClient$call_api(
  url,
  method,
  queryParams = NULL,
  headerParams = NULL,
  body,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgAPIClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
