% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_crop_measurements.R, R/ag_expsite.R
\docType{class}
\name{AgExpSite}
\alias{AgExpSite}
\title{ExpSite}
\format{
R6 class

R6 class
}
\description{
subclass for representing the crop measurements data. It inherents from AgAPI client class 
all the basic parameters and methods.

subclass for representing the experiment-site data. It inherents from AgAPI client class 
all the basic parameters and methods.
}
\section{Super class}{
\code{\link[ragapi:AgAPIClient]{ragapi::AgAPIClient}} -> \code{AgExpSite}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{authentication}}{specifies the password, defaults to NULL}

\item{\code{token}}{specifies the password, defaults to NULL}

\item{\code{user_agent}}{specifies the password, defaults to NULL}

\item{\code{serverURL}}{specifies the password, defaults to 
http://research.cip.cgiar.org/agrofims/api/dev/}

\item{\code{version}}{specifies the version, defaults to 0051}

\item{\code{endPoint}}{specifies the end point value,   tableName/method.php}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{authentication}}{specifies the password, defaults to NULL}

\item{\code{token}}{specifies the password, defaults to NULL}

\item{\code{user_agent}}{specifies the password, defaults to NULL}

\item{\code{serverURL}}{specifies the password, defaults to 
http://research.cip.cgiar.org/agrofims/api/dev/}

\item{\code{version}}{specifies the version, defaults to 0051}

\item{\code{endPoint}}{specifies the end point value,   tableName/method.php}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AgExpSite$new()}}
\item \href{#method-ag_get_expsite_studyId}{\code{AgExpSite$ag_get_expsite_studyId()}}
\item \href{#method-ag_get_sitedesc_studyId}{\code{AgExpSite$ag_get_sitedesc_studyId()}}
\item \href{#method-ag_get_cropsite_studyId}{\code{AgExpSite$ag_get_cropsite_studyId()}}
\item \href{#method-clone}{\code{AgExpSite$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ragapi" data-topic="AgAPIClient" data-id="call_api">}\href{../../ragapi/html/AgAPIClient.html#method-call_api}{\code{ragapi::AgAPIClient$call_api()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ragapi" data-topic="AgAPIClient" data-id="print">}\href{../../ragapi/html/AgAPIClient.html#method-print}{\code{ragapi::AgAPIClient$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgExpSite$new(
  user = NULL,
  password = NULL,
  authentication = NULL,
  token = NULL,
  user_agent = NULL,
  serverURL,
  version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ag_get_expsite_studyId"></a>}}
\if{latex}{\out{\hypertarget{method-ag_get_expsite_studyId}{}}}
\subsection{Method \code{ag_get_expsite_studyId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgExpSite$ag_get_expsite_studyId(
  studyDbId = NULL,
  format = c("json", "list", "data.frame"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ag_get_sitedesc_studyId"></a>}}
\if{latex}{\out{\hypertarget{method-ag_get_sitedesc_studyId}{}}}
\subsection{Method \code{ag_get_sitedesc_studyId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgExpSite$ag_get_sitedesc_studyId(
  studyDbId = NULL,
  format = c("json", "list", "data.frame"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ag_get_cropsite_studyId"></a>}}
\if{latex}{\out{\hypertarget{method-ag_get_cropsite_studyId}{}}}
\subsection{Method \code{ag_get_cropsite_studyId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgExpSite$ag_get_cropsite_studyId(
  studyDbId = NULL,
  format = c("json", "list", "data.frame"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgExpSite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[ragapi:AgAPIClient]{ragapi::AgAPIClient}} -> \code{AgExpSite}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{authentication}}{specifies the password, defaults to NULL}

\item{\code{token}}{specifies the password, defaults to NULL}

\item{\code{user_agent}}{specifies the password, defaults to NULL}

\item{\code{serverURL}}{specifies the password, defaults to 
http://research.cip.cgiar.org/agrofims/api/dev/}

\item{\code{version}}{specifies the version, defaults to 0051}

\item{\code{endPoint}}{specifies the end point value,   tableName/method.php}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{authentication}}{specifies the password, defaults to NULL}

\item{\code{token}}{specifies the password, defaults to NULL}

\item{\code{user_agent}}{specifies the password, defaults to NULL}

\item{\code{serverURL}}{specifies the password, defaults to 
http://research.cip.cgiar.org/agrofims/api/dev/}

\item{\code{version}}{specifies the version, defaults to 0051}

\item{\code{endPoint}}{specifies the end point value,   tableName/method.php}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AgExpSite$new()}}
\item \href{#method-ag_get_expsite_studyId}{\code{AgExpSite$ag_get_expsite_studyId()}}
\item \href{#method-ag_get_sitedesc_studyId}{\code{AgExpSite$ag_get_sitedesc_studyId()}}
\item \href{#method-ag_get_cropsite_studyId}{\code{AgExpSite$ag_get_cropsite_studyId()}}
\item \href{#method-clone}{\code{AgExpSite$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ragapi" data-topic="AgAPIClient" data-id="call_api">}\href{../../ragapi/html/AgAPIClient.html#method-call_api}{\code{ragapi::AgAPIClient$call_api()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ragapi" data-topic="AgAPIClient" data-id="print">}\href{../../ragapi/html/AgAPIClient.html#method-print}{\code{ragapi::AgAPIClient$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgExpSite$new(
  user = NULL,
  password = NULL,
  authentication = NULL,
  token = NULL,
  user_agent = NULL,
  serverURL,
  version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ag_get_expsite_studyId"></a>}}
\if{latex}{\out{\hypertarget{method-ag_get_expsite_studyId}{}}}
\subsection{Method \code{ag_get_expsite_studyId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgExpSite$ag_get_expsite_studyId(
  studyDbId = NULL,
  format = c("json", "list", "data.frame"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ag_get_sitedesc_studyId"></a>}}
\if{latex}{\out{\hypertarget{method-ag_get_sitedesc_studyId}{}}}
\subsection{Method \code{ag_get_sitedesc_studyId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgExpSite$ag_get_sitedesc_studyId(
  studyDbId = NULL,
  format = c("json", "list", "data.frame"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ag_get_cropsite_studyId"></a>}}
\if{latex}{\out{\hypertarget{method-ag_get_cropsite_studyId}{}}}
\subsection{Method \code{ag_get_cropsite_studyId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgExpSite$ag_get_cropsite_studyId(
  studyDbId = NULL,
  format = c("json", "list", "data.frame"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgExpSite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
