% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_fertilizer_measurements.R
\docType{class}
\name{AgFertilizerMea}
\alias{AgFertilizerMea}
\title{AgFertilizerMea}
\format{
R6 class
}
\description{
Class to represent fertilizer measurements data from AgroFIMS. It inherents from AgAPI client class 
all the basic parameters and methods.
}
\section{Super class}{
\code{\link[ragapi:AgAPIClient]{ragapi::AgAPIClient}} -> \code{AgFertilizerMea}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{authentication}}{specifies the password, defaults to NULL}

\item{\code{token}}{specifies the password, defaults to NULL}

\item{\code{user_agent}}{specifies the password, defaults to NULL}

\item{\code{serverURL}}{specifies the password, defaults to 
http://research.cip.cgiar.org/agrofims/api/dev/}

\item{\code{version}}{specifies the version, defaults to 0051}

\item{\code{endPoint}}{specifies the end point value,   tableName/method.php}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{user}}{specifies the user name, defaults to NULL}

\item{\code{password}}{specifies the password, defaults to NULL}

\item{\code{authentication}}{specifies the password, defaults to NULL}

\item{\code{token}}{specifies the password, defaults to NULL}

\item{\code{user_agent}}{specifies the password, defaults to NULL}

\item{\code{serverURL}}{specifies the password, defaults to 
http://research.cip.cgiar.org/agrofims/api/dev/}

\item{\code{version}}{specifies the version, defaults to 0051}

\item{\code{endPoint}}{specifies the end point value,   tableName/method.php}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AgFertilizerMea$new()}}
\item \href{#method-ag_get_fertmea_sitecropId}{\code{AgFertilizerMea$ag_get_fertmea_sitecropId()}}
\item \href{#method-ag_get_fertmea_expsiteId}{\code{AgFertilizerMea$ag_get_fertmea_expsiteId()}}
\item \href{#method-clone}{\code{AgFertilizerMea$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ragapi" data-topic="AgAPIClient" data-id="call_api">}\href{../../ragapi/html/AgAPIClient.html#method-call_api}{\code{ragapi::AgAPIClient$call_api()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ragapi" data-topic="AgAPIClient" data-id="print">}\href{../../ragapi/html/AgAPIClient.html#method-print}{\code{ragapi::AgAPIClient$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgFertilizerMea$new(
  user = NULL,
  password = NULL,
  authentication = NULL,
  token = NULL,
  user_agent = NULL,
  serverURL,
  version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ag_get_fertmea_sitecropId"></a>}}
\if{latex}{\out{\hypertarget{method-ag_get_fertmea_sitecropId}{}}}
\subsection{Method \code{ag_get_fertmea_sitecropId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgFertilizerMea$ag_get_fertmea_sitecropId(siteCropDbId, format, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ag_get_fertmea_expsiteId"></a>}}
\if{latex}{\out{\hypertarget{method-ag_get_fertmea_expsiteId}{}}}
\subsection{Method \code{ag_get_fertmea_expsiteId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgFertilizerMea$ag_get_fertmea_expsiteId(
  expsiteDbId = NULL,
  format = c("json", "list", "data.frame"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgFertilizerMea$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
