% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_data_frame_agapi}
\alias{as_data_frame_agapi}
\title{AgExpDetails as.data.frame}
\usage{
as_data_frame_agapi(.data, tranpose = FALSE)
}
\arguments{
\item{object}{AgExpDetails.}

\item{...}{additional arguments.}
}
\value{
data.frame of the response.
}
\description{
Coerce AgExpDetails responses to data.frame
}
\section{R6 Usage}{
 $as_data_frame_agapi(.data)
}

\seealso{
\code{\link{AgExpDetails}}
}
