% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_data_frame_agexpdetails}
\alias{as_data_frame_agexpdetails}
\title{Convert HTTP responses to AgExpDetails class data}
\usage{
as_data_frame_agexpdetails(.data)
}
\description{
The GET method retrieve a list of values (list data structure), some of them with NULL values,
should be converted to data frame structure.
}
