% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_list_null}
\alias{check_list_null}
\title{Check null values in a list}
\usage{
check_list_null(x)
}
\arguments{
\item{x}{list or list of lists}
}
\description{
Check if there are nulls in the a list or list of lists
}
\author{
Omar Benites
}
