% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_handler}
\alias{error_handler}
\title{Error handling in API's responses}
\usage{
error_handler(x)
}
\description{

}
\references{
See reference in https://github.com/ropensci/ckanr/blob/c5d5bc32be76d3b6447807adc1501803aeeaefa3/R/zzz.R
}
