% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{has_internet}
\alias{has_internet}
\title{agapi_canInternet}
\usage{
has_internet(url = "www.google.com.pe")
}
\arguments{
\item{url}{character; A url to check. Default is \href{https://www.google.com}{www.google.com}.}
}
\value{
NULL when internet access is available or ERROR when internet access is not available
}
\description{
Check for internet connection.
}
\references{
\href{http://stackoverflow.com/questions/5076593/how-to-determine-if-you-have-an-internet-connection-in-r}{Stack Overflow} and \code{\link[curl]{has_internet}}
}
\author{
Omar Benites
}
\concept{agapi_helpers}
